chore(admin): update panels
