+++ src/app/(panel)/dashboard/page.tsx
@@ -14,8 +14,9 @@ const DashboardPage = () => {
             <div className="mb-6">
                 <h1 className="text-4xl font-extrabold leading-tight">Sistema de Evaluación Docente</h1>
                 <p className="mt-2 text-lg text-muted-foreground">
-                    Bienvenido al sistema de evaluación docente para el programa de gerencia de proyectos en postgrados de la
-                    Universidad El Bosque.
+                    Bienvenido al Sistema de Información para la Evaluación Docente, basado en el modelo de evaluación 360 grados
+                    para los programas de posgrado (especializaciones y maestrías) de la Facultad de Ingeniería en Gerencia de
+                    Proyectos de la Universidad El Bosque.
                 </p>
             </div>
         </section>
