+++ src/app/(panel)/dashboard/reports/page.tsx
@@ -10,36 +10,12 @@ import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle }
 import { Download, FileText, Save } from "lucide-react"
 import { getProfessors } from "@/services/professors"
 import type { ProfessorService, SubjectService } from "@/lib/@types/services"
-import { getSubjectsByProfessorId } from "@/services/subjects"
-import { jsPDF } from "jspdf"
+import { getSubjects, getSubjectsByProfessorId } from "@/services/subjects"
 import { createReport, getReports } from "@/services/report"
+import { generateNewReportPDF, generateSavedReportPDF } from "@/lib/report"
+import type { Report } from "@/lib/@types/reports"
 
-interface Report {
-    id: string
-    title: string
-    professor_id: string | null
-    subject_id: string | null
-    evaluation_criteria?: string | null
-    analysis?: string | null
-    comments?: string | null
-    recommendations?: string | null
-    created_at: string
-    professor_name?: string | null
-    subject_name?: string | null
-    professor?: {
-        id: string
-        first_name: string
-        last_name: string
-        email?: string
-    }
-    subject?: {
-        id: string
-        name: string
-        description?: string
-    }
-}
-
-interface ReportState {
+export interface ReportState {
     title: string
     professor: string
     subject: string
@@ -49,16 +25,16 @@ interface ReportState {
     [key: string]: any
 }
 
-const evaluationCriteria = [
-    { id: "teachingQuality", name: "Calidad de Enseñanza" },
-    { id: "communication", name: "Comunicación" },
-    { id: "availability", name: "Disponibilidad" },
-    { id: "fairness", name: "Imparcialidad" },
-    { id: "knowledge", name: "Conocimiento de la Materia" },
-    { id: "organization", name: "Organización del Curso" },
-]
-
 const timeframes = [{ id: "all", name: "Todo el Tiempo" }]
+
+const initialReportState: ReportState = {
+    title: "",
+    professor: "",
+    subject: "",
+    timeframe: "all",
+    comments: "",
+    recommendations: "",
+}
 
 const AdminReportsPage = () => {
     const [activeTab, setActiveTab] = useState("new")
@@ -66,13 +42,7 @@ const AdminReportsPage = () => {
     const [professors, setProfessors] = useState<ProfessorService[]>([])
     const [savedReports, setSavedReports] = useState<Report[]>([])
     const [isLoading, setIsLoading] = useState(true)
-    const [report, setReport] = useState<ReportState>({
-        title: "",
-        professor: "",
-        subject: "",
-        comments: "",
-        recommendations: "",
-    })
+    const [report, setReport] = useState<ReportState>(initialReportState)
 
     const handleChange = (key: keyof ReportState, value: any) => {
         setReport((prev) => ({
@@ -82,13 +52,7 @@ const AdminReportsPage = () => {
     }
 
     const resetForm = () => {
-        setReport({
-            title: "",
-            professor: "",
-            subject: "",
-            comments: "",
-            recommendations: "",
-        })
+        setReport(initialReportState)
         setSubjects([])
     }
 
@@ -134,150 +98,19 @@ const AdminReportsPage = () => {
         }
     }
 
-    const generateNewReportPDF = () => {
-        if (!report.title || !report.professor || !report.subject) {
-            alert("Complete título, profesor y materia antes de generar el PDF")
-            return
-        }
-
-        const doc = new jsPDF()
-        const marginLeft = 20
-        let y = 20
-
-        const professor = professors.find((p) => p.id === report.professor)
-        const subject = subjects.find((s) => s.id === report.subject)
-        const currentDate = new Date().toLocaleDateString("es-ES", {
-            year: "numeric",
-            month: "long",
-            day: "numeric",
-        })
-
-        doc.setFillColor(30, 41, 59)
-        doc.rect(0, 0, 210, 30, "F")
-        doc.setFontSize(20)
-        doc.setTextColor(255, 255, 255)
-        doc.setFont("helvetica", "bold")
-        doc.text("REPORTE DOCENTE", 105, 20, { align: "center" })
-
-        y = 40
-        doc.setTextColor(0, 0, 0)
-        doc.setFontSize(12)
-        doc.setFont("helvetica", "normal")
-        doc.text(`Docente: ${professor?.first_name} ${professor?.last_name}`, marginLeft, y)
-        y += 8
-        doc.text(`Materia: ${subject?.name || "N/A"}`, marginLeft, y)
-        y += 8
-        doc.text(`Fecha: ${currentDate}`, marginLeft, y)
-        y += 10
-
-        doc.setDrawColor(200, 200, 200)
-        doc.line(marginLeft, y, 190, y)
-        y += 10
-
-        doc.setFontSize(14)
-        doc.setFont("helvetica", "bold")
-        doc.text("Análisis y Comentarios", marginLeft, y)
-        y += 8
-        doc.setFontSize(11)
-        doc.setFont("helvetica", "normal")
-        const commentLines = doc.splitTextToSize(report.comments || "No se registraron comentarios.", 170)
-        doc.text(commentLines, marginLeft, y)
-        y += commentLines.length * 6 + 10
-
-        doc.setFontSize(14)
-        doc.setFont("helvetica", "bold")
-        doc.text("Recomendaciones", marginLeft, y)
-        y += 8
-        doc.setFontSize(11)
-        doc.setFont("helvetica", "normal")
-        const recLines = doc.splitTextToSize(report.recommendations || "No se registraron recomendaciones.", 170)
-        doc.text(recLines, marginLeft, y)
-        y += recLines.length * 6 + 10
-
-        doc.setDrawColor(180, 180, 180)
-        doc.line(20, 285, 190, 285)
-        doc.setFontSize(10)
-        doc.setTextColor(100, 100, 100)
-        doc.text("Sistema de Evaluacion Docente Programas Postgrados Gerencia de Proyectos - Universidad El bosque", 105, 290, {
-            align: "center",
-        })
-
-        doc.save(`Reporte_Docente_${professor?.last_name}_${currentDate.replace(/ /g, "_")}.pdf`)
-    }
-
-    const generateSavedReportPDF = (reportId: string) => {
-        const savedReport = savedReports.find((r) => r.id === reportId)
-        if (!savedReport) return
-
-        const doc = new jsPDF()
-        const marginLeft = 20
-        let y = 20
-
-        const currentDate = new Date(savedReport.created_at).toLocaleDateString("es-ES", {
-            year: "numeric",
-            month: "long",
-            day: "numeric",
-        })
-
-        doc.setFillColor(30, 41, 59)
-        doc.rect(0, 0, 210, 30, "F")
-        doc.setFontSize(20)
-        doc.setTextColor(255, 255, 255)
-        doc.setFont("helvetica", "bold")
-        doc.text("REPORTE DOCENTE HISTÓRICO", 105, 20, { align: "center" })
-
-        y = 40
-        doc.setTextColor(0, 0, 0)
-        doc.setFontSize(12)
-        doc.setFont("helvetica", "normal")
-        doc.text(`Docente: ${savedReport.professor_name || "No disponible"}`, marginLeft, y)
-        y += 8
-        doc.text(`Materia: ${savedReport.subject_name || "No disponible"}`, marginLeft, y)
-        y += 8
-        doc.text(`Fecha del informe: ${currentDate}`, marginLeft, y)
-        y += 10
-
-        doc.setDrawColor(200, 200, 200)
-        doc.line(marginLeft, y, 190, y)
-        y += 10
-
-        doc.setFontSize(14)
-        doc.setFont("helvetica", "bold")
-        doc.text("Resumen del Reporte", marginLeft, y)
-        y += 8
-        doc.setFontSize(11)
-        doc.setFont("helvetica", "normal")
-
-        const summary = [
-            `Título: ${savedReport.title}`,
-            `Periodo evaluado: ${currentDate}`,
-            `Materia: ${savedReport.subject_name}`,
-            "Comentarios:",
-            ...doc.splitTextToSize(savedReport.comments || "No se registraron comentarios.", 170),
-            "",
-            "Recomendaciones:",
-            ...doc.splitTextToSize(savedReport.recommendations || "No se registraron recomendaciones.", 170),
-        ]
-
-        doc.text(summary, marginLeft, y)
-
-        doc.setDrawColor(180, 180, 180)
-        doc.line(20, 285, 190, 285)
-        doc.setFontSize(10)
-        doc.setTextColor(100, 100, 100)
-        doc.text("Sistema de Evaluacion Docente Programas Postgrados Gerencia de Proyectos - Universidad El bosque", 105, 290, {
-            align: "center",
-        })
-
-        doc.save(`Reporte_Historico_${savedReport.professor_name}_${currentDate.replace(/ /g, "_")}.pdf`)
-    }
-
     useEffect(() => {
         const loadInitialData = async () => {
             setIsLoading(true)
             try {
                 const [professorsData, reportsData] = await Promise.all([getProfessors(), getReports()])
-                setProfessors(professorsData)
+                setProfessors([
+                    ...professorsData,
+                    {
+                        id: "all",
+                        first_name: "Todos",
+                        last_name: "los Profesores",
+                    } as ProfessorService,
+                ])
                 setSavedReports(
                     reportsData.filter(
                         (r) =>
@@ -306,8 +139,13 @@ const AdminReportsPage = () => {
             }
 
             try {
-                const data = await getSubjectsByProfessorId(report.professor)
-                setSubjects(data)
+                if (report.professor === "all") {
+                    const data = await getSubjects()
+                    setSubjects(data)
+                } else {
+                    const data = await getSubjectsByProfessorId(report.professor)
+                    setSubjects(data)
+                }
             } catch (error) {
                 console.error("Error al cargar las materias:", error)
                 setSubjects([])
@@ -433,7 +271,7 @@ const AdminReportsPage = () => {
                                     <Save className="mr-2 h-4 w-4" />
                                     Guardar Borrador
                                 </Button>
-                                <Button onClick={generateNewReportPDF}>
+                                <Button onClick={() => generateNewReportPDF(report, professors, subjects)}>
                                     <FileText className="mr-2 h-4 w-4" />
                                     Generar PDF
                                 </Button>
@@ -473,7 +311,7 @@ const AdminReportsPage = () => {
                                                             <Button
                                                                 variant="ghost"
                                                                 size="icon"
-                                                                onClick={() => generateSavedReportPDF(report.id)}
+                                                                onClick={() => generateSavedReportPDF(savedReports, report.id)}
                                                                 title="Descargar PDF"
                                                             >
                                                                 <Download className="h-4 w-4" />
